// module_19/src/app/interfaces/borrow.interface.ts
// book (objectId) — Mandatory. References the borrowed book’s ID.
// quantity (number) — Mandatory. Positive integer representing the number of copies borrowed.
// dueDate (date) — Mandatory. The date by which the book must be returned.

import { ObjectId } from "mongoose";

export interface IBorrow {
	book: ObjectId;
	quantity: number;
	dueDate: Date;
}